improvement and clenup
- Delete unused files
- Use TS
- Reorg files structure
